whisper-backend/pyproject.toml
[tool.poetry]
name = "whisper-backend"
version = "0.1.0"
description = ""
authors = ["Veter-Alex <Vet-an@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
